requirements.txt
openai>=1.3.0
requests>=2.28.0
python-dotenv>=0.19.0
Pillow>=9.0.0
beautifulsoup4>=4.11.0 